Fixed client side anim not updated on weapon change.
If a client switches between different weapons, the global anim might match the one from the new weapon. This can cause issues with weapons having different bodygroups.

This fix checks if the client switched to a new weapon and allows HUD_SendWeaponAnim to update with the proper bodygroup if needed.